requirements.txt
flask==3.0.3
werkzeug==3.0.3
pandas==2.1.4
numpy==1.26.4
plotly==5.18.0
google-generativeai==0.5.0
openpyxl==3.1.2
scikit-learn==1.4.2
xgboost==2.0.3
scipy==1.11.4
sqlalchemy==2.0.30
pymysql==1.1.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pyarrow==15.0.0
gunicorn==21.2.0
